+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/worktables/gui/ResultSlot.java
@@ -1,0 +1,22 @@
+package com.codetaylor.mc.artisanworktables.modules.worktables.gui;
+
+import net.minecraft.item.ItemStack;
+import net.minecraftforge.items.IItemHandler;
+import net.minecraftforge.items.SlotItemHandler;
+
+import javax.annotation.Nonnull;
+
+public class ResultSlot
+    extends SlotItemHandler {
+
+  public ResultSlot(IItemHandler itemHandler, int index, int xPosition, int yPosition) {
+
+    super(itemHandler, index, xPosition, yPosition);
+  }
+
+  @Override
+  public boolean isItemValid(@Nonnull ItemStack stack) {
+
+    return false;
+  }
+}